Database and addPlayer added
This app has now a simple database and i already made the method to add new players, seems like its working. At first it crashed because my database attribute in the class was null, i dont fully understand how it works but now it does.
Next thing i have to get the list of all the players to see if they are really being added correctly.